done A-Gradle
