+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 public abstract class Task {
 
     private final String taskname;
